Convert search fragment to activity
