+++ app/src/main/res/values-es/strings.xml
@@ -829,7 +829,7 @@
     <string name="online">Conectado</string>
     <string name="summary_change_24h">%1$s (%2$s) Hoy</string>
     <string name="performance">Performance</string>
-    <string name="token_name_hint">Token Name</string>
+    <string name="token_name_hint">Search token</string>
     <string name="assets">Assets</string>
     <string name="defi_header">DeFi</string>
     <string name="governance_header">Gobernancia</string>
