+++ app/src/main/res/values-fr/strings.xml
@@ -847,7 +847,7 @@
     <string name="online">Connecté</string>
     <string name="summary_change_24h">%1$s (%2$s) Aujourd\'hui</string>
     <string name="performance">Performance</string>
-    <string name="token_name_hint">Token Name</string>
+    <string name="token_name_hint">Search token</string>
     <string name="assets">Assets</string>
     <string name="defi_header">DeFi</string>
     <string name="governance_header">Gouvernance</string>
