+++ app/src/main/res/values-vi/strings.xml
@@ -849,7 +849,7 @@
     <string name="online">Trực tuyến</string>
     <string name="summary_change_24h">%1$s (%2$s) Hôm nay</string>
     <string name="performance">Performance</string>
-    <string name="token_name_hint">Token Name</string>
+    <string name="token_name_hint">Search token</string>
     <string name="assets">Assets</string>
     <string name="defi_header">DeFi</string>
     <string name="governance_header">Quản trị</string>
