+++ app/src/main/res/values-zh/strings.xml
@@ -829,7 +829,7 @@
     <string name="online">在线</string>
     <string name="summary_change_24h">%1$s (%2$s) 今天</string>
     <string name="performance">Performance</string>
-    <string name="token_name_hint">Token Name</string>
+    <string name="token_name_hint">Search token</string>
     <string name="assets">资产</string>
     <string name="defi_header">DeFi</string>
     <string name="governance_header">治理</string>
